Issue #21781, _ssl: Fix asn1obj2py() on Windows 64-bit, "s#" format requires size to be a Py_ssize_t, not an int. _ssl.c is now "Py_ssize_t clean".

+++ Modules/_ssl.c
@@ -3416,7 +3416,7 @@ asn1obj2py(ASN1_OBJECT *obj)
     int nid;
     const char *ln, *sn;
     char buf[100];
-    int buflen;
+    Py_ssize_t buflen;
 
     nid = OBJ_obj2nid(obj);
     if (nid == NID_undef) {
